% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{allFieldStrings}
\alias{allFieldStrings}
\title{List of the Distinct String Values for a Specified Python Field}
\usage{
allFieldStrings(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Given a list of objects, all of which have a string-valued field specified by what,
returns a list of the distinct strings found in all the elements of objects.
}
\section{Python Functions}{

allFieldStrings(objects, what)
}

