% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClasses.R
\docType{class}
\name{Element_Python-class}
\alias{Element_Python}
\alias{Element_Python-class}
\title{Proxy for Python Class Element in Module xml.etree.ElementTree}
\description{
Proxy for Python Class Element in Module xml.etree.ElementTree
}

