% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{getPersonae}
\alias{getPersonae}
\title{Python Dictionary of the Speakers in a List of Speeches}
\usage{
getPersonae(..., .ev = XRPython::RPython(), .get = TRUE)
}
\description{
[Python Documentation]
Returns a list of character strings describing the personae in the play.  In the style of
the XML coding, each string is the name of the person as it will appear in speeches given
by that character, followed by a description.  <NOT YET>:For personae listed in a group rather than
separately, this function inserts the group description, in square brackets.
}
\section{Python Functions}{

getPersonae(play)
}

