% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{getPlay_Python}
\alias{getPlay_Python}
\title{Python Function to Parse XML File for Play}
\usage{
getPlay_Python(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Parse the file name, or if it's not a string, assume it's
a previously parsed tree.  (Should of course check that).
}
\section{Python Functions}{

getPlay(what)
}

