% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{getScenes}
\alias{getScenes}
\title{Python List of all Scenes of a Play}
\usage{
getScenes(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Return a list of the scenes in the XML object "play".  Each element of the list is
an object of class "Scene" with fields "title", "act" and "data"
(the XML element for the scene).  The argument can alternatively be an Act object to obtain a list
of scenes from just that act.
}
\section{Proxy Function}{

getScenes(play) [Python]
}

