% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyClassesExtended.R
\name{installPlays}
\alias{installPlays}
\title{Install Parsed Plays and Speeches Lists}
\usage{
installPlays(what = .playsTable$keys, report = TRUE)
}
\arguments{
\item{what}{character vector identifying the plays to install.  By default installs all the 37 plays (which can take
a minute or more, depending on the hardware).}

\item{report}{if \code{TRUE} (the default) the function will report its progress.}
}
\value{
the vector of keys corresponding to the requested plays.
}
\description{
Installs all the plays whose keys, or uniquely identifying strings from their title, are supplied as \code{what}.
Installing means storing the parsed plays and derived speeches lists in tables for the session.  If the user
has write permission on the installed package's directories, serialized versions will be written there, to be used
in later R sessions, rather than recomputing the parse trees and speech lists.
}
\details{
Use of this function is never required.  If a play is accessed as an object of class \code{\linkS4class{Play}}, the
same information will be stored for the session and written to the package files, if the user has write permission.
It's perhaps less likely to be confusing if the plays are installed initially, as access to the tables is essentially
instantaneous.  Unserializing the saved files may be mildly noticeable as a delay, and the initial parse and computation
is roughly twice as long again.
}

