% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{parseXML}
\alias{parseXML}
\title{Python Function to Parse XML File}
\usage{
parseXML(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
Python Function to Parse XML File
}
\section{Python Functions}{

parse(source, parser =)
}

