% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proxyClassesExtended.R
\docType{class}
\name{Play-class}
\alias{Play}
\alias{Play-class}
\title{A Class for the Python/XML Version of a Play}
\description{
The plays are parsed from the set of XML files into Python \code{"ElementTree"} objects.
An R object from class \code{"Play"} has a proxy to the parsed play plus fields for the
names of the personae and a proxy to a Python list of all the speeches.
This class extends the proxy class for \code{"ElementTree"}.
}
\section{Fields}{

\describe{
\item{\code{personae}}{Character vector of the descriptions of the personae in the play.  Note that this
comes from the play itself.  The names of speakers usually match one of the personae in the \code{grep()}
sense, but not always.}

\item{\code{title}}{The character string title, as found in the XML representation.}

\item{\code{speeches}}{Proxy for a Python listof all the speeches (each an object
of Python class \code{"Speech"}).  This is precomputed when the \code{"Play"} object is initialized;
the speeches list tends to be input to many of the interesting analyses.  If you want to suppress
precomputation, explicitly set this field to a Python list, as in the example below.}

\item{\code{key}}{The character string identifying the play in the table and also the name of the original XML file.}
}}
\section{Methods}{

\describe{
\item{\code{find(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: find(path, namespaces =)}

\item{\code{findall(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: findall(path, namespaces =)}

\item{\code{findtext(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: findtext(path, default =, namespaces =)}

\item{\code{getiterator(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: getiterator(tag =)}

\item{\code{getroot(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: getroot()}

\item{\code{iter(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: iter(tag =)}

\item{\code{iterfind(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: iterfind(path, namespaces =)}

\item{\code{parse(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: parse(source, parser =)}

\item{\code{write(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: write(file_or_filename, encoding =, xml_declaration =, default_namespace =, method =)}

\item{\code{write_c14n(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: write_c14n(file)}
}}

