% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playDates.R
\name{playDates}
\alias{playDates}
\title{Estimated Dates for Specified Plays}
\usage{
playDates(text)
}
\arguments{
\item{text}{character vector. Either the key(s) for the plays or some text that uniquely matches the corresponding title.  Call
\code{\link{playTitles}()} to see the table of keys and titles.}
}
\value{
the years as a character vector.
}
\description{
The vector \code{text} specifies one or more plays; used internally these will be the keys identifying the
various plays.  Otherwise they should match the title of the desired play, as for \code{\link{findPlay}()}.
}
\details{
The dates are taken from those given on opensourceshakespeare.org.  The actual dates of first performance of the plays
is not always known, and some are debated.
Our philosophy is just to regard these as useful calibration when looking for style changes or other
indicators of the changing environment in which the plays appeared.
}

