% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlays.R
\name{playTitles}
\alias{playTitles}
\title{Titles of Shakespeare's Plays and Corresponding Keys}
\usage{
playTitles()
}
\description{
This function returns, as  data frame, the titles of Shakespeare's plays as stored in XML files and the corresponding keys used to identify
the individual plays.  Either the key or a unique string matching a title can be given as the name to \code{\link{getPlay}()}.
}

