% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proxyClassesExtended.R
\name{printSpeeches}
\alias{printSpeeches}
\title{Print a List of Speeches or Speech Fragments}
\usage{
printSpeeches(speeches, printSeparator = TRUE)
}
\arguments{
\item{speeches}{A list of speeches, either a list in Python or the result of getting
such an object back to R.}

\item{printSeparator}{Should a special separator line be printed between speeches.  Generally you would want this if the list was a result of a search, but not
to print a contiguous part of the play, e.g., a scene.}
}
\description{
Given a list of speech objects, print them all.  The list may be a portion
of a play (e.g., a scene) or a list of speech fragments generated by
\code{\link{searchSpeeches}()} and \code{\link{speechFragments}()}.
}

