% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{searchSpeeches}
\alias{searchSpeeches}
\title{List of Lines Matching Specified Text in a List of Speeches}
\usage{
searchSpeeches(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Given a character string, text, and a list of speeches, returns a parallel list
each element of which is the list of matching lines in the speech, as returned by the
findText() method.  The arguments token and ignoreCase are passed to that method.
If token is True, the match will be against the word tokens constructed for the speech list when
the corresponding play was intialized.  Otherwise the match is against the text of the lines.
The default for argument ignoreCase is True if tokens is True else False.
}
\section{Python Functions}{

searchSpeeches(text, speeches, tokens =, ignoreCase =)
}

