\name{shakespeare-package}
\alias{shakespeare-package}
\alias{shakespeare}
\docType{package}
\title{
Tools for Text Analysis of Shakespeare Plays
}
\description{
XML files holding the plays can be parsed via Python tools
}
\details{
\tabular{ll}{
Package: \tab shakespeare\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-16\cr
License: \tab What license is it under?\cr
}
}
\author{
A Shakespeare Fan
}
\keyword{ package }

