% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{speakers}
\alias{speakers}
\title{Python List of the Persons Listed for the Play}
\usage{
speakers(..., .ev = XRPython::RPython(), .get = TRUE)
}
\description{
[Python Documentation]
A dictionary whose keys are all the names of speakers with speeches in the list.  The entry will be
the total character count of the speeches if argument count is True, else just True.
The list of speeches can come from the "speeches" field of a Play object or as the result of calling
getSpeeches() for a whole play, an act or a scene.
}
\section{Proxy Function}{

speakers(speeches, count =) [Python]
}

