% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{speechFragments}
\alias{speechFragments}
\title{List of Speeches Constructed with Fragments}
\usage{
speechFragments(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Given some a list of speeches and a parallel list of matches to lines within each speech, returns a
constructed list of speech fragments for any of the speeches that have nonempty matches.  All matched lines will
be included plus some preceding and following lines as specified by the before=
and after= arguments.
The list of matched lines will typically come from a call to searchSpeeches().
}
\section{Python Functions}{

speechFragments(speeches, matches, before =, after =, filler =)
}

