% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyClassesExtended.R
\docType{class}
\name{SpeechList-class}
\alias{SpeechList}
\alias{SpeechList-class}
\title{Class for Python List of Speech objects}
\description{
The R reference class \code{"SpeechList"} extends the proxy class for a Python list.  The assertion (unchecked) is
that the elements of the Python list are of class \code{"Speech"}.
In particular, every object of R class \code{\linkS4class{Play}} has a field containing a list of all the speeches
in the play, computed when the play is installed for this session.
}
\section{Methods}{

\describe{
\item{\code{print()}}{this method prints the entire list of speeches; e.g., the whole play, in contrast to the automatic show() method, which prints only a summary.}
}}

