% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{speechSearch}
\alias{speechSearch}
\title{List of Speech Fragments with Lines Matching Specified Text}
\usage{
speechSearch(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Given some text and a list of speeches, returns a
constructed list of speech fragments for any of the speeches that contains the
text (as determined by the hasText() method.  All lines containing the text will
be included plus some preceding and following lines as specified by the before=
and after= arguments.
(Argument emph= is not currently used.)
}
\section{Proxy Function}{

speechSearch(text, speeches, before =, after =, emph =, filler =) [Python]
}

