% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/shakespeare.R
\name{speechTokens}
\alias{speechTokens}
\title{Vectors of Tokens for all Speeches According to Speaker}
\usage{
speechTokens(speeches, .get = NA)
}
\description{
Given a proxy for a play, act or scene object in Python, or for a list of
speeches, the function returns a named list, with names being the speakers.
The corresponding element is a list of
tokens as produced by the \code{word_tokenizer()} function in Python, converting the
speeches by this speaker into a stream of tokens.
}
\details{
This function extends the proxy for a Python function of the same name
and argument.  If the optional argument \code{.get} is \code{TRUE},
the value returned by the Python function is converted and
the lines for each speech are unlisted into a character vector.
}
\author{
John M. Chambers
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
}
\seealso{
tokens This returns the same information but in one vector for the tokens in all speeches.
}
\keyword{Programming}

