% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{speechTokens_Python}
\alias{speechTokens_Python}
\title{A Python Dictionary of all Speeches by Speaker}
\usage{
speechTokens_Python(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
A dictionary whose keys are the names of all speakers with speeches in the list.
The corresponding element is a list of all the speeches spoken by that speaker, with
each speech converted into a list of its tokens.
}
\details{
The argument can be from the "speeches" field of a Play object or the result
of any other computation.
The argument could also be an Act, Scene or Play:  any object for which getSpeeches()
returns a list of speeches.
}
\section{Proxy Function}{

speechTokens(speeches) [Python]
}

