% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Proxy_toEtree_Function.R
\name{toEtree}
\alias{toEtree}
\title{Convert an Object to a Parsed Etree}
\usage{
toEtree(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Parse the file name, or if it's not a string, assume it's
a previously parsed tree.  (Should of course check that).
}
\section{Proxy Function}{

toEtree(what) [Python]
}

