% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shakespeare.R
\name{tokens}
\alias{tokens}
\title{Contents of Speeches Tokenized and Split According to Speaker}
\usage{
tokens(speeches)
}
\arguments{
\item{speeches}{Proxy for a Python object containing the relevant speeches.
This may be a list of speeches or and object for which \code{getSpeeches()}
returns such a list.  Possiblities include a Play, Act or Scene object.}
}
\description{
The Python NLTK software is used to produce tokenized versions of all
the speeches supplied in the argument.
The resulting dictionary is returned to R as a named list.
The list corresponding to a particular speaker is unlisted and
the token \code{"$"} used as a line separator is replaced with new line
characters.
}

