% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{tokens_Python}
\alias{tokens_Python}
\title{Speeches Tokenized in a Python Dictionary by Speaker}
\usage{
tokens_Python(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
A dictionary whose keys are all the names of speakers with speeches in the list.
The corresponding element is a list of all the tokens spoken by each speaker.
Special tokens '$' and '/>' are inserted to mark the end of lines and the end of individual
speeches.
}
\details{
The argument can be from the "speeches" field of a Play object or the result
of any other computation.
The argument could also be an Act, Scene or Play:  any object for which getSpeeches()
returns a list of speeches.
}
\section{Python Functions}{

tokens(speeches)
}

