% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{wordsUsed}
\alias{wordsUsed}
\title{Python List of the Distinct Words in a List of Tokens}
\usage{
wordsUsed(..., .ev = XR::getInterface(), .get = NA)
}
\description{
[Python Documentation]
Given a list of tokens, returns a list of the distinct words included.
Words are converted to lower case for comparison.
Common words are excluded by default, being defined as the stopwords set from
NLTK supplemented with a few common words in Elizabethan English.
Punctuation is also excluded by default.  Optional second & third arguments in
the call can override if supplied as True.
}
\section{Proxy Function}{

wordsUsed(tokens, includeCommon =, includePunctuation =) [Python]
}

