% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClasses.R
\docType{class}
\name{Element_Python-class}
\alias{Element_Python}
\alias{Element_Python-class}
\title{Proxy for Python Class Element in Module xml.etree.ElementTree}
\description{
Proxy for Python Class Element in Module xml.etree.ElementTree
}
\section{Methods}{

\describe{
\item{\code{append(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: append(element)}

\item{\code{clear(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: clear()}

\item{\code{extend(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: extend(elements)}

\item{\code{find(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: find(path, namespaces =)}

\item{\code{findall(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: findall(path, namespaces =)}

\item{\code{findtext(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: findtext(path, default =, namespaces =)}

\item{\code{get(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: get(key, default =)}

\item{\code{getchildren(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: getchildren()}

\item{\code{getiterator(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: getiterator(tag =)}

\item{\code{insert(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: insert(index, element)}

\item{\code{items(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: items()}

\item{\code{iter(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: iter(tag =)}

\item{\code{iterfind(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: iterfind(path, namespaces =)}

\item{\code{itertext(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: itertext()}

\item{\code{keys(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: keys()}

\item{\code{makeelement(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: makeelement(tag, attrib)}

\item{\code{remove(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: remove(element)}

\item{\code{set(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: set(key, value)}
}}

