% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ProxyClasses.R
\docType{class}
\name{ElementTree_Python-class}
\alias{ElementTree_Python}
\alias{ElementTree_Python-class}
\title{Proxy for Python Class ElementTree in Module xml.etree.ElementTree}
\description{
Proxy for Python Class ElementTree in Module xml.etree.ElementTree
}
\section{Methods}{

\describe{
\item{\code{find(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: find(path, namespaces =)}

\item{\code{findall(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: findall(path, namespaces =)}

\item{\code{findtext(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: findtext(path, default =, namespaces =)}

\item{\code{getiterator(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: getiterator(tag =)}

\item{\code{getroot(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: getroot()}

\item{\code{iter(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: iter(tag =)}

\item{\code{iterfind(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: iterfind(path, namespaces =)}

\item{\code{parse(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: parse(source, parser =)}

\item{\code{write(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: write(file_or_filename, encoding =, xml_declaration =, default_namespace =, method =)}

\item{\code{write_c14n(..., .ev = XRPython::RPython(), .get = NA)}}{Python Method: write_c14n(file)}
}}

