% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlClasses.R
\docType{class}
\name{ElementTree_Python-class}
\alias{ElementTree_Python}
\alias{ElementTree_Python-class}
\title{Proxy for Python Class ElementTree in Module xml.etree.ElementTree}
\description{
Proxy for Python Class ElementTree in Module xml.etree.ElementTree
}

