% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{getActs}
\alias{getActs}
\title{List (in Python) of the Acts of a Play}
\usage{
getActs(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Return a list of the acts in the XML object "play".  Each element of the list is
an object of class "Act" with fields "title" and "data" (the XML element for the act).
}
\section{Proxy Function}{

getActs(play) [Python]
}

