% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{getPlay}
\alias{getPlay}
\title{Get the Python Tree for a Play}
\usage{
getPlay(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
Parse the file name, or if it's not a string, assume it's
a previously parsed tree.  (Should of course check that).
}
\section{Proxy Function}{

getPlay(what) [Python]
}

