% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{speakers}
\alias{speakers}
\title{Python List of the Persons Listed for the Play}
\usage{
speakers(..., .ev = XRPython::RPython(), .get = TRUE)
}
\description{
[Python Documentation]
A list of all the speakers found in the list. The argument
can be a list of speeches or an object (Play, Act, Scene) for which
getSpeeches() returns such a list.
}
\section{Proxy Function}{

speakers(speeches) [Python]
}

