% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyFunctions.R
\name{tokens_Python}
\alias{tokens_Python}
\title{Speeches Tokenized in a Python Dictionary by Speaker}
\usage{
tokens_Python(..., .ev = XRPython::RPython(), .get = NA)
}
\description{
[Python Documentation]
A dictionary whose keys are all the names of speakers with speeches in the list.
The corresponding element is a list of all the tokens spoken by each speaker.
}
\details{
The argument can be from the "speeches" field of a Play object or the result
of any other computation.
The argument could also be an Act, Scene or Play:  any object for which getSpeeches()
returns a list of speeches.
}
\section{Proxy Function}{

tokens(speeches) [Python]
}

